+++ src/profile/mfa/mfa_factor_sources/off_device_mnemonic_factor_source/mod.rs
@@ -1,5 +1,7 @@
+mod off_device_factor_source_hint;
 mod off_device_mnemonic_factor_source;
 mod off_device_mnemonic_factor_source_uniffi_fn;
 
+pub use off_device_factor_source_hint::*;
 pub use off_device_mnemonic_factor_source::*;
 pub use off_device_mnemonic_factor_source_uniffi_fn::*;
